requirements.txt
absl-py==0.12.0
attrs==20.3.0
gin-config==0.4.0
imgaug==0.4.0
matplotlib==3.3.4
nibabel==3.2.1
numpy==1.19.5
pandas==1.2.3
pdfcombine==1.1.4
scipy==1.6.1
tensorflow==2.4.1
tqdm==4.59.0
